Elevation: steps must be an integer

+++ elevation_controller_server.py
@@ -61,7 +61,7 @@ class ElevationController:
     def move_to(self, percentage_amplitude):
         if percentage_amplitude < 0 or percentage_amplitude > 1:
             raise ValueError("Percentage of amplitude must be between 0 and 1: " + str(percentage_amplitude))
-        steps = (percentage_amplitude - self.position()) * self.amplitude()
+        steps = int((percentage_amplitude - self.position()) * self.amplitude())
         direction = self.UP_DIRECTION if steps < 0 else self.DOWN_DIRECTION
         self.__stepper.move(direction, abs(steps))
         return True
